Added MQTT logic

+++ Net.Bluewalk.NukiBridge2Mqtt/Models/Lock.cs
@@ -5,7 +5,7 @@ namespace Net.Bluewalk.NukiBridge2Mqtt.Models
     public class Lock
     {
         [SerializeAs(Name = "nukiId")]
-        public long NukiId { get; set; }
+        public int NukiId { get; set; }
 
         [SerializeAs(Name = "name")]
         public string Name { get; set; }
